+++ src/asp/Core/IpMatchingAlgs.cc
@@ -17,7 +17,8 @@
 
 #include <asp/Core/IpMatchingAlgs.h>         // Lightweight header
 #include <vw/InterestPoint/InterestData.h>
-
+#include <vw/InterestPoint/Matcher.h>
+#include <boost/filesystem.hpp>
 using namespace vw;
 
 namespace asp {
@@ -48,4 +49,33 @@ void align_ip(vw::TransformPtr const& tx_left,
   return;
 } // End align_ip
 
+// Heuristics for where to load the match file from  
+std::string match_filename(std::string const& clean_match_files_prefix,
+                           std::string const& match_files_prefix,
+                           std::string const& out_prefix,
+                           std::string const& image1_path,
+                           std::string const& image2_path) {
+  
+  if (clean_match_files_prefix != "") {
+
+    std::string match_file = vw::ip::clean_match_filename(clean_match_files_prefix, image1_path,
+                                                          image2_path);
+    if (!boost::filesystem::exists(match_file)) 
+      vw_throw(ArgumentErr() << "Missing IP file: " << match_file);
+    return match_file;
+    
+  } else if (match_files_prefix != "") {
+    
+    std::string match_file = vw::ip::match_filename(match_files_prefix, image1_path,
+                                                    image2_path);
+    if (!boost::filesystem::exists(match_file)) 
+      vw_throw(ArgumentErr() << "Missing IP file: " << match_file);
+    return match_file;
+    
+  }
+  
+  return vw::ip::match_filename(out_prefix, image1_path,
+                                image2_path);
+}
+  
 } // end namespace asp